+++ src/fields/PlainText.php
@@ -16,6 +16,7 @@ use craft\base\MergeableFieldInterface;
 use craft\base\SortableFieldInterface;
 use craft\elements\Entry;
 use craft\fields\conditions\TextFieldConditionRule;
+use craft\helpers\Html;
 use craft\helpers\StringHelper;
 
 /**
@@ -249,6 +250,22 @@ class PlainText extends Field implements InlineEditableFieldInterface, SortableF
         return TextFieldConditionRule::class;
     }
 
+    /**
+     * @inheritdoc
+     */
+    public function getPreviewHtml(mixed $value, ?ElementInterface $element = null): string
+    {
+        $previewHtml = parent::getPreviewHtml($value, $element);
+
+        if (!$this->code) {
+            return $previewHtml;
+        }
+
+        return Html::tag('div', $previewHtml, [
+            'class' => 'code',
+        ]);
+    }
+    
     /**
      * @inheritdoc
      */