+++ src/services/ElementSources.php
@@ -153,6 +153,27 @@ class ElementSources extends Component
     }
 
     /**
+     * Returns whether the given source exists.
+     *
+     * @param class-string<ElementInterface> $elementType The element type class
+     * @param string $sourceKey The source key
+     * @param string $context The context
+     * @param bool $withDisabled Whether disabled sources should be included
+     * @return bool
+     * @since 5.7.11
+     */
+    public function sourceExists(string $elementType, string $sourceKey, string $context = self::CONTEXT_INDEX, bool $withDisabled = false): bool
+    {
+        foreach ($this->getSources($elementType, $context, $withDisabled) as $source) {
+            if (($source['key'] ?? null) === $sourceKey) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
      * Returns whether the given custom source should be available for the current user.
      *
      * @param array $source